reorganizing repo to hub multiple projects. project till now is placed under products_and_orders_app
