+++ products_and_orders_app/src/main/java/com/jsn/products/productsapp/controller/OrderController.java
@@ -1,0 +1,34 @@
+package com.jsn.products.productsapp.controller;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.jsn.products.productsapp.model.Order;
+import com.jsn.products.productsapp.service.OrderService;
+
+@RestController
+@RequestMapping("/api/orders")
+public class OrderController {
+
+	Logger logger = LogManager.getLogger(OrderController.class);
+
+    @Autowired
+    private OrderService orderService;
+
+    @PostMapping
+    public Order postMethodName(@RequestBody Order order) {
+        return orderService.createOrder(order);
+    }
+
+    @GetMapping("/{id}")
+    public Order getOrderById(@PathVariable Long id) {
+        return orderService.getOrderById(id);
+    }
+}
