+++ products_and_orders_app/src/main/java/com/jsn/products/productsapp/model/Product.java
