+++ products_and_orders_app/src/main/java/com/jsn/products/productsapp/repository/OrderRepository.java
@@ -1,0 +1,9 @@
+package com.jsn.products.productsapp.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.jsn.products.productsapp.model.Order;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+
+}
