--- src/main/java/com/jsn/products/productsapp/kafka/KafkaProducerConfig.java
@@ -1,59 +1,0 @@
-package com.jsn.products.productsapp.kafka;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-import org.springframework.kafka.support.serializer.JsonSerializer;
-
-import com.jsn.products.productsapp.model.Order;
-import com.jsn.products.productsapp.model.Product;
-
-@Configuration
-public class KafkaProducerConfig {
-
-	@Value(value = "${spring.kafka.bootstrap-servers}")
-    private String bootstrapAddress;
-	
-	/*
-	 * knote: specify serializer for key and value of messages. Producer instances are thread safe
-	 * 
-	 */
-	/*
-	 * knote: this class became Order specific, how do i create config to send other types of objects? use string serializer + json object mapper?
-	 */
-    @Bean
-    public ProducerFactory<String, Order> producerFactory() {
-        Map<String, Object> configProps = new HashMap<>();
-        configProps.put(
-          ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, 
-          bootstrapAddress);
-        configProps.put(
-          ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, 
-          StringSerializer.class);
-        configProps.put(
-          ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, 
-          JsonSerializer.class);
-        return new DefaultKafkaProducerFactory<>(configProps);
-    }
-
-    /*
-     * knote: KakfaTemplate instances are also thread safe, and use of one instance is recommended
-     
-    @Bean
-    public KafkaTemplate<String, String> kafkaTemplate() {
-        return new KafkaTemplate<>(producerFactory());
-    }
-    */
-    @Bean
-    public KafkaTemplate<String, Order> kafkaTemplate() {
-        return new KafkaTemplate<>(producerFactory());
-    }
-}
