--- src/main/java/com/jsn/products/productsapp/repository/OrderRepository.java
@@ -1,9 +1,0 @@
-package com.jsn.products.productsapp.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.jsn.products.productsapp.model.Order;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-
-}
